Remove <form /> elements from README and tests
At this point in the curriculum, fellows have not yet been taught about event listeners or how and why to use `event.preventDefault()` with a form submit.
In its current form, the lab asks fellows to create forms, submit them, and then add new elements to the page based on user input. This will not work without calling `event.preventDefault()` as the page will reload, and this will also cause trouble for automation.

Similar to https://github.com/joinpursuit/combining-html-and-javascript-plus-dom/pull/66, I am suggesting that we remove forms from the lab, using only inputs and buttons, with corresponding updates to the tests. This will allow fellows to solve the problems with the tools and concepts they have been given so far.